+++ src/app/pretrips/page.tsx
@@ -1,7 +1,8 @@
 import { Suspense } from 'react'
-import AddTrip from '@/components/ui/trips/add-trip/addTrip'
-import DisplayTrips from '@/components/ui/trips/display-trips/DisplayTrips'
-import TripsTable from '@/components/ui/trips/display-trips/trips-table'
+import DisplayTripsFallback from '@/lib/components/ui/fallbacks/display-trips-fallback'
+import AddTrip from '@/lib/components/ui/trips/add-trip/addTrip'
+import DisplayTrips from '@/lib/components/ui/trips/display-trips/DisplayTrips'
+import TripsTable from '@/lib/components/ui/trips/display-trips/trips-table'
 import { fetchCurrentTrips, fetchPastTrips } from '@/lib/DAL/trips'
 
 export default function Page() {
@@ -10,14 +11,14 @@ export default function Page() {
 	return (
 		<div className="flex flex-col items-center justify-items-center gap-4 p-4">
 			<AddTrip />
-			<Suspense>
+			<Suspense fallback={<DisplayTripsFallback />}>
 				<DisplayTrips
 					currentTripsComponent={
 						<TripsTable tripsPromise={currentTripsPromise} />
 					}
 					pastTripsComponent={
 						<TripsTable tripsPromise={pastTripsPromise} />
-					}
+					}					
 				/>
 			</Suspense>
 		</div>
