+++ src/lib/components/ui/trips/display-trips/trips-table.tsx
